+++ src/java/ManagerQueries/RichestRep.java
@@ -52,10 +52,10 @@ public class RichestRep extends HttpServlet {
                 session.setAttribute("userBean", userBean);
             }
             
-            Integer ssn = userBean.getSocialSecurityNumber();
+            Long ssn = userBean.getSocialSecurityNumber();
             
             //should set the thingy
-            Integer customerSSN = Integer.parseInt(request.getParameter("customerSSN"));
+            Long customerSSN = Long.parseLong(request.getParameter("customerSSN"));
             //should set the thingy
             String stockSymbol = request.getParameter("stockSymbol");
